feat(group-field-change): change behavior of changing group county.
