+++ server/src/DBService/LandingPage/Mutation.ts
@@ -8,14 +8,6 @@ export const CHANGE_DESK_ID = gql`
   }
 `;
 
-export const UPDATE_COUNTY_BY_GROUP_ID = gql`
-mutation updateInvestigatorByGroupId($county: Int!, $selectedGroups: [UUID!]!) {
-  updateCountyByGroupId(input: {county: $county, selectedGroups: $selectedGroups}) {
-    clientMutationId
-  }
-}
-`;
-
 export const UPDATE_DESK_BY_GROUP_ID = gql`
 mutation updateInvestigatorByGroupId($desk: Int!, $selectedGroups: [UUID!]!) {
   updateDeskByGroupId(input: {desk: $desk, selectedGroups: $selectedGroups}) {
